Fix cumulative variance overflow on 32-bit integer platforms
Use an explicit float64 dtype for cumulative time indices before squaring them
in the TBR variance formula. This prevents platform-dependent integer overflow
for long test periods, keeping cumulative variance calculations stable across
macOS, Linux, and Windows.

+++ src/tbr/core/effects.py
@@ -161,7 +161,7 @@ def calculate_cumulative_variance(
         raise ValueError("test_x_values cannot be empty")
 
     n = len(test_x_values)
-    T_values = np.arange(1, n + 1)  # [1, 2, 3, ..., n]
+    T_values = np.arange(1, n + 1, dtype=np.float64)  # [1, 2, 3, ..., n]
 
     # Calculate cumulative means efficiently using vectorized operations
     cumsum_x = np.cumsum(test_x_values)
